+++ packages/attention/index.js
@@ -1,7 +1,8 @@
-import { css, html, LitElement } from 'lit'
-import { ifDefined } from 'lit/directives/if-defined.js'
-import { classes, kebabCaseAttributes, generateRandomId } from '../utils'
-import { attention as ccAttention } from '@warp-ds/css/component-classes'
+import { css, html } from 'lit';
+import WarpElement from '@warp-ds/elements-core';
+import { ifDefined } from 'lit/directives/if-defined.js';
+import { classes, kebabCaseAttributes, generateRandomId } from '../utils';
+import { attention as ccAttention } from '@warp-ds/css/component-classes';
 import {
   opposites,
   rotation,
@@ -13,7 +14,7 @@ import { messages as nbMessages } from './locales/nb/messages.mjs'
 import { messages as fiMessages } from './locales/fi/messages.mjs'
 import { activateI18n } from '../i18n'
 
-class WarpAttention extends kebabCaseAttributes(LitElement) {
+class WarpAttention extends kebabCaseAttributes(WarpElement) {
   static properties = {
     // Whether Attention element should be visible.
     show: { type: Boolean, reflect: true },
@@ -26,13 +27,15 @@ class WarpAttention extends kebabCaseAttributes(LitElement) {
     callout: { type: Boolean, reflect: true },
     // Whether Attention element is rendered as a popover
     popover: { type: Boolean, reflect: true },
+    // Whether Attention element is rendered as a highlight
+    highlight: { type: Boolean, reflect: true },
     // Render Attention element without an arrow
     noArrow: { type: Boolean, reflect: true },
   }
 
-  static styles = [
+  static styles = [WarpElement.styles,
     css`
-      @unocss-placeholder #attention {
+      #attention {
         position: absolute;
         z-index: 50;
         visibility: var(--attention-visibility);
@@ -58,6 +61,7 @@ class WarpAttention extends kebabCaseAttributes(LitElement) {
     this.tooltip = false
     this.callout = false
     this.popover = false
+    this.highlight = false;
     this.noArrow = false
   }
 
@@ -176,6 +180,13 @@ class WarpAttention extends kebabCaseAttributes(LitElement) {
           comment:
             'Default screenreader message for popover speech bubble in the attention component',
         })
+      case this.highlight:
+        return i18n._({
+          id: 'attention.aria.highlight',
+          message: 'highlighted speech bubble',
+          comment:
+            'Default screenreader message for highlighted speech bubble in the attention component',
+        })
       default:
         return ''
     }
@@ -193,6 +204,23 @@ class WarpAttention extends kebabCaseAttributes(LitElement) {
       this._targetEl.setAttribute('aria-details', attentionMessageId)
     }
   }
+
+
+  get _activeVariantClasses() {
+    const variantProps = {
+      callout: this.callout,
+      popover: this.popover,
+      tooltip: this.tooltip,
+      highlight: this.highlight
+    }
+
+    const activeVariant = Object.keys(variantProps).find(b => !!variantProps[b]) || '';
+
+    return {
+      wrapper: ccAttention[activeVariant],
+      arrow: ccAttention[`arrow${activeVariant.charAt(0).toUpperCase() + activeVariant.slice(1)}`]
+    }
+  };
 
   firstUpdated() {
     this.setAriaLabels()
@@ -222,10 +250,8 @@ class WarpAttention extends kebabCaseAttributes(LitElement) {
   get _wrapperClasses() {
     return classes({
       [ccAttention.base]: true,
-      [ccAttention.tooltip]: this.tooltip,
-      [ccAttention.callout]: this.callout,
-      [ccAttention.popover]: this.popover,
-    })
+      [this._activeVariantClasses.wrapper]: true
+    });
   }
 
   get _arrowClasses() {
@@ -237,10 +263,8 @@ class WarpAttention extends kebabCaseAttributes(LitElement) {
           this._arrowDirection.slice(1)
         }`
       ]]: true,
-      [ccAttention.arrowTooltip]: this.tooltip,
-      [ccAttention.arrowCallout]: this.callout,
-      [ccAttention.arrowPopover]: this.popover,
-    })
+      [this._activeVariantClasses.arrow]: true
+    });
   }
 
   get _arrowHtml() {