+++ packages/badge/index.js
@@ -1,22 +1,32 @@
-import { css, html, LitElement } from 'lit';
-import { classes } from '../utils';
-import { badge as ccBadge } from '@warp-ds/css/component-classes';
+import { html } from "lit";
+import WarpElement from "@warp-ds/elements-core";
+import { classes } from "../utils";
+import { badge as ccBadge } from "@warp-ds/css/component-classes";
 
-class WarpBadge extends LitElement {
+class WarpBadge extends WarpElement {
   static properties = {
-    variant: { type: 'neutral' | 'info' | 'positive' | 'warning' |'negative' | 'disabled' | 'notification' | 'price' },
-    position: { type: 'top-left' | 'top-right' | 'bottom-right' | 'bottom-left' },
+    variant: {
+      type:
+        "neutral" |
+        "info" |
+        "positive" |
+        "warning" |
+        "negative" |
+        "disabled" |
+        "notification" |
+        "price",
+    },
+    position: {
+      type: "top-left" | "top-right" | "bottom-right" | "bottom-left",
+    },
   };
 
-  static styles =
-    css`
-      @unocss-placeholder
-      `;
+  static styles = [WarpElement.styles];
 
   constructor() {
     super();
 
-    this.variant = 'neutral';
+    this.variant = "neutral";
   }
 
   get _class() {
@@ -24,10 +34,10 @@ class WarpBadge extends LitElement {
       [ccBadge.base]: true,
       [ccBadge[this.variant]]: true,
       [ccBadge.positionBase]: !!this.position,
-      [ccBadge.positionTL]: this.position === 'top-left',
-      [ccBadge.positionTR]: this.position === 'top-right',
-      [ccBadge.positionBR]: this.position === 'bottom-right',
-      [ccBadge.positionBL]: this.position === 'bottom-left',
+      [ccBadge.positionTL]: this.position === "top-left",
+      [ccBadge.positionTR]: this.position === "top-right",
+      [ccBadge.positionBR]: this.position === "bottom-right",
+      [ccBadge.positionBL]: this.position === "bottom-left",
     });
   }
 
@@ -40,8 +50,8 @@ class WarpBadge extends LitElement {
   }
 }
 
-if (!customElements.get('w-badge')) {
-  customElements.define('w-badge', WarpBadge);
+if (!customElements.get("w-badge")) {
+  customElements.define("w-badge", WarpBadge);
 }
 
 export { WarpBadge };